another one

+++ 92_div2c.cpp
@@ -11,7 +11,7 @@ int main()
     {
         w[s[i]-'a']++;
     }
-    long long int n = strlen(s),c=n-1;
+    long long int n =strlen(s),c=n-1;
     d[0] = 1;
     for(long long int i=n/2+1;i<=n;i++) 
     {
@@ -32,7 +32,7 @@ int main()
     {
         cout << "YES\n" ;
         char v = max_element(w,w+26)-w +'a';
-        for(int i=0;i<n;i++)
+        for(long long int i=0;i<n;i++)
         {
             if(d[i]==0)
             { 
@@ -40,7 +40,7 @@ int main()
                 w[v-'a']--;
             }
         }
-        for(int i=0;i<n;i++)
+        for(long long int i=0;i<n;i++)
         {
             if(d[i]==1) 
             {
